refactor: chatbot 파일 구조 재정리(#47)
