+++ src/routes/Chatbot/hooks/useChatbot.js
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
-import { getChat, postChat } from '../api/chatBotApi';
+import { getChat, postChat } from '../../../api/chatBotApi';
 import { useCookies } from 'react-cookie';
 
 function useChatbot() {
